Use AGP support for Kotlin sources when it is available
This change introduces support for Kotlin source directories
when they specified in AGP. Please note this this feature is
not yet available in the released AGP, but adding suport in
the KGP is the first step. If the support is not available,
old behavior is preserved.

The location of Kotlin sources is (as known by KGP) is passed
to AGP, in order to sync the two models.