[Debugger] Support Local Variables in evaluate expressions
Reuse the CodegenInfo from the old backend which contains a mapping
from descriptors of captured variables to a positional argument of the
method that wraps the evaluated fragment.

This commit introduces an injected interceptor that redirects accesses
to the symbol table for local variables and maps them to the
appropriate parameter instead of looking for an actual local
variable. The injection is crude but works, we'll see how far it
carries this first effort.

This commit copies the first few layers of the generator stack in
PSI2IR to account for the fact that we don't have source for the
module down to the body of the method we're generating. This should
probably ebe refactored.