Refactor Comparison Propagation for Extension Receivers
Relnote:"""
This change impacts a few aspects of how "comparison propagation" happens in composable functions which may have observable behavior differences.

The essence of the changes are:

1. Unused parameters don't affect skipping

If parameters of a composable function aren't used in the body of the function, then they are no longer used to determine whether or not the function can be "skipped" or not. Prior to this change, this was the case for the "extension receiver" parameter of a function and also the "dispatch receiver", but now it is the case for _all_ parameters. This is unlikely to break any behavior that was being relied upon in practical code, but it is slightly more likely that this could break tests which assert on a given function being "called", and potentially ignore parameters being passed in. This will also provide a performance improvement globally since fewer comparisons will be taking place in general.

2. Composables with required params that might be stable still can skip

Prior to this change, if a required parameter to a composable function was not provably "Stable", then the function would never skip, even if a stable version of that parameter was passed in to it. This has now been changed so that the code required to "skip" a composable function will be generated if the parameter could _possibly_ be stable, so that if the parameter _is_ stable, it will skip. This will be a performance improvement in some cases as more functions will possibly skip, and has the potential to be a slight regression in other cases as the runtime will be performing more comparisons. This will also slightly impact bytecode size since more functions will have the machinery to skip them generated. This behavior was the original/intended behavior of stable inference.

3. Extension receiver slot index is now always 0

Prior to this change, composable lambdas with "extension receiver" parameters were incorrectly determining whether or not they were skippable. This was because the bitmask which is used to coordinate comparison state/metadata of parameters for a composable lambda with n formal parameters and 1 extension parameter was using the 0th index for the extension parameter in some cases, and the nth index in other places, and similarly, using the 1st index for the 1st formal parameter in some places, and the 0th index for that parameter in other cases. This misalignment was causing some composable lambdas to incorrectly skip when they weren't supposed to be skipping, as well as causing them to sometimes _execute_ when it was possible to skip them. This is a breaking change but will only impact composable lambdas with an extension receiver and Unit return type. It is recommended that because of this change, people be careful using compose code which has been compiled with versions of the compose compiler both before and after this change.
"""

Fixes: 174778953 ( https://issuetracker.google.com/issues/174778953 )
Bug: 175717335 ( https://issuetracker.google.com/issues/175717335 )

Change-Id: I42e2b911bc50f4b8eaa8a8b0e5ac72d95f3bfc34 ( https://android-review.googlesource.com/q/I42e2b911bc50f4b8eaa8a8b0e5ac72d95f3bfc34 )

Moved from: https://github.com/androidx/androidx/commit/77282c908eecf4b2b977bb65775747783919bb46