Fix lost task dependency on resolved-metadata-provider tasks
The task dependency might have been lost when the metadata provider
returned empty metadata dependencies set. In that case, the compile
classpath did not include any file collection built-by the metadata
provider task, and the task (doing nothing) was not included in the
task graph.

This could lead to confusing behavior. Fix that by always adding a
FileCollection built-by the metadata provider task.