[Kotlin/Native] Reduce the overhead of EnterFrame/LeaveFrame by a factor of 4x on Windows
TlsGetValue/TlsSetValue and TlsAlloc has less overhead than the thread local implementation
that the compiler provides on windows since it is a direct call to the implementation.
Since EnterFrame/LeaveFrame is used everywhere this leads to noticeable improvements,
still slower than linux and mac, but in the same order of magnitude.