Remove unnecessary group realization redudancy check.
A check for redundant realization of groups was removed as
it is unnecessary and complicated to avoid. Changing this to be
idempotent that treats multiple realizations of child gorups as
redundant and ignored instead of an internal error.

Fixes: b/182682230 ( https://issuetracker.google.com/issues/182682230 ), b/186924463 ( https://issuetracker.google.com/issues/186924463 )
Test: ./gradlew :composer:c:c-h:i-t:tDUT
Change-Id: If67c412a14ae985e551b335cef9cc34d0f1b1b07 ( https://android-review.googlesource.com/q/If67c412a14ae985e551b335cef9cc34d0f1b1b07 )

Moved from: https://github.com/androidx/androidx/commit/3c005c4f8349930c8dd97bd1aafdea95a83a8dfd