Annotate composables in frontend instead of backend
Add annotations to the descriptors of lambdas which are inferred composable
in the frontend.  Delete the IrComposableAnnotator which was annotating in
the backend.  We still need to figure out if this is kosher especially as
the FIR becomes more of a thing (this may be an abuse of an extension
point that I added previously, so we may need to revert this CL in the
future), but it would be nice if annotating in the frontend is a viable
long-term way of passing information from the frontend to backend.

Change-Id: Ib5a71a35eef2d4e581201823dcddc72bbf3fba30 ( https://android-review.googlesource.com/q/Ib5a71a35eef2d4e581201823dcddc72bbf3fba30 )

Moved from: https://github.com/androidx/androidx/commit/193b3a917f09da6c896dd13dc530bf546db06ea2