Make sure that loads of the LiveLiterals class has correct offsets.
The load of the receiver for the call to the getter needs to
also have the offset of the expression replaced. Otherwise, it
will take the line number of whatever is generated before it.
That can lead to debugging issues such as stepping into unreachable
branches.

Bug: 190481816 ( https://issuetracker.google.com/issues/190481816 )
Test: manual debugger test
Change-Id: Id441ac711b9518bb035a9771558985d2c7463f1e ( https://android-review.googlesource.com/q/Id441ac711b9518bb035a9771558985d2c7463f1e )

Moved from: https://github.com/androidx/androidx/commit/ab0c604888ad7def2dc759a5a430146b5455668c