Add workaround for compiler downloader for 1.5.20 release binaries
Because of migration K/N to kotlin repo we still have some problems
with version transformation.
Right now, when in 1.5.20-release branch we have build number
(and DeployVersion) something like 1.5.20-256 K/N mechanic interprets it
as Release version and almost all artifacts have names just 1.5.20.
For example: kotlin-native-prebuilt-linux-1.5.20.tar.gz
But they are stored in the separate folder releases/1.5.20-256
and I use this in this commit.

For release version the build number should be -1, because the version
should be just 1.5.20 but we should check it or even revert the commit