FIR: Initial check-in of diagnostics diff and delta tools.
The diff tool extracts data about the differences in diagnostics tests
for FE1.0 vs FIR. The goal is not to get a 100% match, but it shows gaps
in the implementation of FIR diagnostics.

The delta tool shows the impact of a set of changes in getting FIR
diagnostics closer to FE1.0 diagnostics.