[EE-IR] Fragment Compiler Entrypoint in Ir Backend.
This commit introduces a new entrypoint in the IR Backend, and starts
the work of accomodating the evaluator in PSI2IR.

The evaluator expects a certain class and method structure of the
compiled fragment, but PSI is only supplied for the actual fragment.

So, to that end, this commit introduces a new "front half" of psi2ir
that "synthesizes" the module, class and method structure around the
fragment before calling into the existing psi2ir pipeline to obtain
the IR for the fragment.

The primary complication so far is handling the captured variables of
the fragment: they are mapped to parameters of the method surrounding
the fragment, and passed as arguments on evaluation. Hence, the IR
translation of the fragment needs to remap captured variables to the
appropriate parameter, in essentially all places they can be referred
to in the fragment (and hence in psi2ir). To simplify this, we
introduce a decorated symbol table that intercepts symbol look-ups and
remaps as appropriate.

Other cases that dispatches based on descriptor (see
`CallGenerator.kt`) needs other metadata to generate the correct code.

Finally, in passing we resolve a small leftover from previous
refactoring that left an assertion re. allowing IR to _assign_ to
parameters of methods.