[ee-ir] Handle _host_ declarations as external calls
The "hosting" code - the underlying code in the IDE project that is
being debugged is not passed to the psi2ir pipeline as
`DeserializedContainerSource`.

However, the information is all at hand.

This commit introduces a very scraped shim to intercept calls for
names of source files in order to compute facade classes for top level
declarations in the host code (which are "ir external declarations"
that must be resolved in ExpressionCodegen).

Solves 14 remaining cases:
```
JVM_WITH_OLD_EVALUATOR
1
JVM_WITH_IR_EVALUATOR
51
JVM_IR_WITH_OLD_EVALUATOR
40
JVM_IR_WITH_IR_EVALUATOR
77
```