Split LVT record for known nulls
Since they are not spilled, the logic for splitting LVT records, that
is applied for spilled variables, was not applied for known nulls.
Fix that by applying the logic to them.

Do not duplicate $result in LVT

Merge consequent LVT records that is, if LVT record starts where
previous one ends, merge them.
 #KT-47749 Fixed