Dont't apply cacheRedirectory file for included build
In the first commit of this PR, we are applying the cacheRedirectory
file to all build.gradle files including build files from included build
like buildSrc, which is not equivalent to the old behavior. With this
commit, we dont' apply the file to included build by checking whether
the file path is valid. If the file path is not valid, it means the
current build file is from included build.